clean up connect commands
