+++ internal/cmd/connect/command_test.go
@@ -133,8 +133,7 @@ func (suite *ConnectTestSuite) SetupTest() {
 
 func (suite *ConnectTestSuite) newCmd() *command {
 	prerunner := cliMock.NewPreRunnerMock(&ccloud.Client{Connect: suite.connectMock, Kafka: suite.kafkaMock}, nil, nil, suite.conf)
-	cmd := New(suite.conf, prerunner, suite.analyticsClient)
-	return cmd
+	return New(prerunner, suite.analyticsClient)
 }
 
 func (suite *ConnectTestSuite) TestPauseConnector() {
