+++ internal/cmd/connect/utils.go
@@ -10,26 +10,28 @@ import (
 )
 
 const (
-	Config         = "config"
-	Name           = "name"
-	ConnectorClass = "connector.class"
+	config         = "config"
+	name           = "name"
+	connectorClass = "connector.class"
 )
 
 func getConfig(cmd *cobra.Command) (*map[string]string, error) {
-	fileName, err := cmd.Flags().GetString(Config)
+	fileName, err := cmd.Flags().GetString(config)
 	if err != nil {
 		return nil, errors.Wrap(err, "error reading --config as string")
 	}
-	var options map[string]string
-	options, err = parseConfigFile(fileName)
+
+	options, err := parseConfigFile(fileName)
 	if err != nil {
 		return nil, errors.Wrapf(err, "unable to parse config %s", fileName)
 	}
-	_, nameExists := options[Name]
-	_, classExists := options[ConnectorClass]
+
+	_, nameExists := options[name]
+	_, classExists := options[connectorClass]
 	if !nameExists || !classExists {
 		return nil, errors.Errorf(errors.MissingRequiredConfigsErrorMsg, fileName)
 	}
+
 	return &options, nil
 }
 
@@ -45,27 +47,28 @@ func parseConfigFile(fileName string) (map[string]string, error) {
 	kvPairs := make(map[string]string)
 	var options map[string]interface{}
 
-	err = json.Unmarshal(jsonFile, &options)
-
-	if err != nil {
+	if err := json.Unmarshal(jsonFile, &options); err != nil {
 		return nil, errors.Wrapf(err, errors.ParseConfigErrorMsg, fileName)
 	}
+
 	for key, val := range options {
 		if val2, ok := val.(string); ok {
 			kvPairs[key] = val2
 		} else {
 			// We support object-as-a-value only for "config" key.
-			if key != Config {
-				return nil, errors.Errorf("Only string value is permitted for the configuration : %s", key)
+			if key != config {
+				return nil, errors.Errorf(`only string values are permitted for the configuration "%s"`, key)
 			}
+
 			configMap, ok := val.(map[string]interface{})
 			if !ok {
-				return nil, errors.Errorf("Value for the configuration : %s is malformed", Config)
+				return nil, errors.Errorf(`value for the configuration "%s" is malformed`, config)
 			}
+
 			for configKey, configVal := range configMap {
 				value, isString := configVal.(string)
 				if !isString {
-					return nil, errors.Errorf("Only string value is permitted for the configuration : %s", configKey)
+					return nil, errors.Errorf(`only string values are permitted for the configuration "%s"`, configKey)
 				}
 				kvPairs[configKey] = value
 			}
