+++ test/fixtures/output/connect/connect-plugin-describe.golden
@@ -1,5 +1,5 @@
-Following are the required configs: 
-connector.class: GcsSink
+The following are required configs:
+connector.class : GcsSink
 kafka.api.key : ["kafka.api.key" is required] 
 kafka.api.secret : ["kafka.api.secret" is required] 
 topics : ["topics" is required] 
@@ -8,4 +8,3 @@ gcs.credentials.config : ["gcs.credentials.config" is required]
 gcs.bucket.name : ["gcs.bucket.name" is required] 
 time.interval : ["data.format" is required Value "null" doesn't belong to the property's "time.interval" enum] 
 tasks.max : ["tasks.max" is required] 
-
